Add support for Katar Pro XT

+++ src/daemon/bragi_common.h
@@ -8,7 +8,7 @@
 #include <stdint.h>
 
 long int bragi_get_property(usbdevice* kb, const uchar prop);
-int bragi_set_property(usbdevice* kb, const uchar prop, const uchar val);
+int bragi_set_property(usbdevice* kb, const uchar prop, const ushort val);
 
 #define bragi_check_success(a, b) _bragi_check_success((a), (b), INDEX_OF(kb, keyboard), __func__, __FILE__, __LINE__)
